# lib/api-spec/openapi.orval.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Api",
    "version": "0.2.0",
    "description": "Docuvia — Universal VCS Knowledge Graph Engine API"
  },
  "servers": [{ "url": "/api", "description": "Base API path" }],
...
  "paths": {
...
    "/projects/{id}/export": {
      "get": {
        "operationId": "exportProject",
        "tags": ["projects"],
        "summary": "Export project knowledge base as JSON",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "Exported knowledge base",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ProjectExport" } }
            }
          }
        }
      }
    },
    "/projects/{id}/export/md": {
      "get": {
        "operationId": "exportProjectMarkdown",
        "tags": ["projects"],
        "summary": "Export project knowledge base as Markdown",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "Exported knowledge base as markdown",
            "content": { "text/markdown": { "schema": { "type": "string", "format": "binary" } } }
          }
        }
      }
    },
...
    "/documents/{id}/affiliate": {
      "post": {
        "operationId": "affiliateDocument",
        "tags": ["documents"],
        "summary": "Associate an unaffiliated document with a project",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AffiliateDocumentInput" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated document",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Document" } }
            }
          },
          "404": {
            "description": "Document not found",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Project": {
        "type": "object",
        "required": [
          "id",
          "name",
          "repoUrl",
          "status",
          "l2Count",
          "l3Count",
          "commitCount",
          "createdAt"
        ],
        "properties": {
          "id": { "type": "integer" },
          "name": { "type": "string" },
          "repoUrl": { "type": "string" },
          "description": { "type": "string", "nullable": true },
          "status": { "type": "string", "enum": ["active", "indexing", "paused", "error"] },
          "l2Count": { "type": "integer" },
          "l3Count": { "type": "integer" },
          "commitCount": { "type": "integer" },
          "createdAt": { "type": "string" },
          "updatedAt": { "type": "string" },
          "vcsType": { "type": "string", "enum": ["git", "svn"], "default": "git" },
          "svnUrl": { "type": "string", "nullable": true }
        }
      },
...
      "ProjectExport": {
        "type": "object",
        "required": ["project", "l1Tags", "l2Nodes", "l3Nodes", "commits", "exportedAt"],
        "properties": {
          "project": { "$ref": "#/components/schemas/Project" },
          "l1Tags": { "type": "array", "items": { "$ref": "#/components/schemas/L1Tag" } },
          "l2Nodes": { "type": "array", "items": { "$ref": "#/components/schemas/L2Node" } },
          "l3Nodes": { "type": "array", "items": { "$ref": "#/components/schemas/L3Node" } },
          "commits": { "type": "array", "items": { "$ref": "#/components/schemas/Commit" } },
          "exportedAt": { "type": "string" }
        }
      },
      "L1Tag": {
        "type": "object",
        "required": ["id", "name", "category", "usageCount", "createdAt"],
        "properties": {
          "id": { "type": "integer" },
          "name": { "type": "string" },
          "category": { "type": "string" },
          "description": { "type": "string", "nullable": true },
          "isAnchored": { "type": "boolean" },
          "usageCount": { "type": "integer" },
          "createdAt": { "type": "string" }
        }
      },
...
      "L2Node": {
        "type": "object",
        "required": [
          "id",
          "projectId",
          "name",
          "type",
          "l3Count",
          "l1TagIds",
          "aiGenerated",
          "createdAt"
        ],
        "properties": {
          "id": { "type": "integer" },
          "projectId": { "type": "integer" },
          "name": { "type": "string" },
          "type": { "type": "string", "enum": ["package", "module", "pcd", "sys-uncategorized"] },
          "description": { "type": "string", "nullable": true },
          "l3Count": { "type": "integer" },
          "l1TagIds": { "type": "array", "items": { "type": "integer" } },
          "aiGenerated": { "type": "boolean" },
          "needsReview": { "type": "boolean" },
          "isSystem": { "type": "boolean" },
          "isBootstrapConfirmed": { "type": "boolean" },
          "pathPatterns": { "type": "array", "items": { "type": "string" } },
          "createdAt": { "type": "string" }
        }
      },
...
      "L3Node": {
        "type": "object",
        "required": ["id", "l2NodeId", "title", "nodeType", "aiGenerated", "createdAt"],
        "properties": {
          "id": { "type": "integer" },
          "l2NodeId": { "type": "integer" },
          "title": { "type": "string" },
          "content": { "type": "string", "nullable": true },
          "nodeType": { "type": "string", "enum": ["change", "rule", "decision", "context"] },
          "commitHash": { "type": "string", "nullable": true },
          "aiGenerated": { "type": "boolean" },
          "confidence": { "type": "number", "nullable": true },
          "noiseScore": { "type": "number", "nullable": true },
          "createdAt": { "type": "string" }
        }
      },
...
      "Commit": {
        "type": "object",
        "required": ["id", "projectId", "hash", "message", "author", "valid", "createdAt"],
        "properties": {
          "id": { "type": "integer" },
          "projectId": { "type": "integer" },
          "hash": { "type": "string" },
          "message": { "type": "string" },
          "author": { "type": "string" },
          "valid": { "type": "boolean" },
          "filterScore": { "type": "number", "nullable": true },
          "l2NodeId": { "type": "integer", "nullable": true },
          "createdAt": { "type": "string" }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": ["error"],
        "properties": { "error": { "type": "string" } }
      },
...
      "Document": {
        "type": "object",
        "required": ["id", "l2NodeId", "filename", "docType", "createdAt"],
        "properties": {
          "id": { "type": "integer" },
          "projectId": { "type": "integer", "nullable": true },
          "l2NodeId": { "type": "integer" },
          "filename": { "type": "string" },
          "docType": {
            "type": "string",
            "enum": ["markdown", "txt", "pdf", "docx", "pptx", "build_artifact"]
          },
          "content": { "type": "string" },
          "contentHash": {
            "type": "string",
            "description": "SHA-256 hash of raw content",
            "nullable": true
          },
          "commitSha": {
            "type": "string",
            "description": "Optional commit SHA to associate with this document",
            "nullable": true
          },
          "affiliatedAt": {
            "type": "string",
            "description": "ISO-8601 timestamp when document was associated with a project",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": ["unaffiliated", "affiliated"],
            "default": "unaffiliated"
          },
          "createdAt": { "type": "string" }
        }
      },
...
      "AffiliateDocumentInput": {
        "type": "object",
        "required": ["projectId"],
        "properties": { "projectId": { "type": "integer" } }
      }
    }
  }
}